Add mysql connector and move existing database code
