+++ homecontrol/database/sqlite/connection.py
