+++ homecontrol/database/sqlite/database.py
@@ -1,5 +1,5 @@
 import re
-from homecontrol.database.connection import DatabaseConnection
+from homecontrol.database.sqlite.connection import DatabaseConnection
 from homecontrol.scheduling.structs import SchedulerDatabaseConfig
