+++ requirements.txt
@@ -13,6 +13,8 @@ itsdangerous==2.1.2
 Jinja2==3.1.2
 MarkupSafe==2.1.2
 msmart==0.2.5
+mysql-connector-python==8.0.33
+protobuf==3.20.3
 pycryptodome==3.18.0
 pytz==2023.3
 requests==2.31.0
