+++ scripts/get_monitoring.py
@@ -1,8 +1,9 @@
 from datetime import datetime
 from typing import List
+
 from homecontrol.api.monitoring import TempDataPoint
 from homecontrol.client.client import Client
-from homecontrol.database.database import Database
+from homecontrol.database.sqlite.database import Database
 
 
 def get_all_room_data():
